+++ src/SdfEntityCreator.cc
@@ -21,6 +21,7 @@
 #include "ignition/gazebo/Events.hh"
 #include "ignition/gazebo/SdfEntityCreator.hh"
 
+#include "ignition/gazebo/components/AirPressureSensor.hh"
 #include "ignition/gazebo/components/Altimeter.hh"
 #include "ignition/gazebo/components/AngularVelocity.hh"
 #include "ignition/gazebo/components/Camera.hh"
@@ -37,6 +38,7 @@
 #include "ignition/gazebo/components/Joint.hh"
 #include "ignition/gazebo/components/JointAxis.hh"
 #include "ignition/gazebo/components/JointType.hh"
+#include "ignition/gazebo/components/Lidar.hh"
 #include "ignition/gazebo/components/Light.hh"
 #include "ignition/gazebo/components/LinearAcceleration.hh"
 #include "ignition/gazebo/components/LinearVelocity.hh"
@@ -55,6 +57,7 @@
 #include "ignition/gazebo/components/Static.hh"
 #include "ignition/gazebo/components/ThreadPitch.hh"
 #include "ignition/gazebo/components/Visual.hh"
+#include "ignition/gazebo/components/WindMode.hh"
 #include "ignition/gazebo/components/World.hh"
 
 class ignition::gazebo::SdfEntityCreatorPrivate
@@ -172,6 +175,8 @@ Entity SdfEntityCreator::CreateEntities(const sdf::Model *_model)
       components::Name(_model->Name()));
   this->dataPtr->ecm->CreateComponent(modelEntity,
       components::Static(_model->Static()));
+  this->dataPtr->ecm->CreateComponent(
+      modelEntity, components::WindMode(_model->EnableWind()));
 
   // NOTE: Pose components of links, visuals, and collisions are expressed in
   // the parent frame until we get frames working.
@@ -188,6 +193,13 @@ Entity SdfEntityCreator::CreateEntities(const sdf::Model *_model)
     {
       this->dataPtr->ecm->CreateComponent(linkEntity,
           components::CanonicalLink());
+    }
+
+    // Set wind mode if the link didn't override it
+    if (!this->dataPtr->ecm->Component<components::WindMode>(linkEntity))
+    {
+      this->dataPtr->ecm->CreateComponent(
+          linkEntity, components::WindMode(_model->EnableWind()));
     }
   }
 
@@ -242,6 +254,12 @@ Entity SdfEntityCreator::CreateEntities(const sdf::Link *_link)
       components::Name(_link->Name()));
   this->dataPtr->ecm->CreateComponent(linkEntity,
       components::Inertial(_link->Inertial()));
+
+  if (_link->EnableWind())
+  {
+    this->dataPtr->ecm->CreateComponent(
+        linkEntity, components::WindMode(_link->EnableWind()));
+  }
 
   // Visuals
   for (uint64_t visualIndex = 0; visualIndex < _link->VisualCount();
@@ -403,31 +421,40 @@ Entity SdfEntityCreator::CreateEntities(const sdf::Sensor *_sensor)
 
   if (_sensor->Type() == sdf::SensorType::CAMERA)
   {
-    auto elem = _sensor->Element();
-
     this->dataPtr->ecm->CreateComponent(sensorEntity,
-        components::Camera(elem));
+        components::Camera(*_sensor));
   }
   else if (_sensor->Type() == sdf::SensorType::GPU_LIDAR)
   {
-    auto elem = _sensor->Element();
-
     this->dataPtr->ecm->CreateComponent(sensorEntity,
-        components::GpuLidar(elem));
+        components::GpuLidar(*_sensor));
+  }
+  else if (_sensor->Type() == sdf::SensorType::LIDAR)
+  {
+    // \todo(anyone) Implement CPU-base lidar
+    // this->dataPtr->ecm->CreateComponent(sensorEntity,
+    //     components::Lidar(*_sensor));
+    ignwarn << "Sensor type LIDAR not supported yet. Try using"
+      << "a GPU LIDAR instead." << std::endl;
   }
   else if (_sensor->Type() == sdf::SensorType::DEPTH_CAMERA)
   {
-    auto elem = _sensor->Element();
-
     this->dataPtr->ecm->CreateComponent(sensorEntity,
-        components::DepthCamera(elem));
+        components::DepthCamera(*_sensor));
+  }
+  else if (_sensor->Type() == sdf::SensorType::AIR_PRESSURE)
+  {
+    this->dataPtr->ecm->CreateComponent(sensorEntity,
+        components::AirPressureSensor(*_sensor));
+
+    // create components to be filled by physics
+    this->dataPtr->ecm->CreateComponent(sensorEntity,
+        components::WorldPose(math::Pose3d::Zero));
   }
   else if (_sensor->Type() == sdf::SensorType::ALTIMETER)
   {
-     auto elem = _sensor->Element();
-
     this->dataPtr->ecm->CreateComponent(sensorEntity,
-        components::Altimeter(elem));
+        components::Altimeter(*_sensor));
 
     // create components to be filled by physics
     this->dataPtr->ecm->CreateComponent(sensorEntity,
@@ -437,10 +464,8 @@ Entity SdfEntityCreator::CreateEntities(const sdf::Sensor *_sensor)
   }
   else if (_sensor->Type() == sdf::SensorType::IMU)
   {
-    auto elem = _sensor->Element();
-
     this->dataPtr->ecm->CreateComponent(sensorEntity,
-            components::Imu(elem));
+            components::Imu(*_sensor));
 
     // create components to be filled by physics
     this->dataPtr->ecm->CreateComponent(sensorEntity,