+++ src/systems/physics/Physics.cc
@@ -68,6 +68,7 @@
 // Components
 #include "ignition/gazebo/components/AngularAcceleration.hh"
 #include "ignition/gazebo/components/AngularVelocity.hh"
+#include "ignition/gazebo/components/BatterySoC.hh"
 #include "ignition/gazebo/components/CanonicalLink.hh"
 #include "ignition/gazebo/components/ChildLinkName.hh"
 #include "ignition/gazebo/components/Collision.hh"
@@ -80,6 +81,7 @@
 #include "ignition/gazebo/components/JointPosition.hh"
 #include "ignition/gazebo/components/JointType.hh"
 #include "ignition/gazebo/components/JointVelocity.hh"
+#include "ignition/gazebo/components/JointVelocityCmd.hh"
 #include "ignition/gazebo/components/LinearAcceleration.hh"
 #include "ignition/gazebo/components/LinearVelocity.hh"
 #include "ignition/gazebo/components/Link.hh"
@@ -197,9 +199,13 @@ class ignition::gazebo::systems::PhysicsPrivate
   /// in the ECM. This is the reverse map of entityCollisionMap.
   public: std::unordered_map<ShapePtrType, Entity> collisionEntityMap;
 
-  /// \brief a map between joint entity ids in the ECM to Joint Entities in
+  /// \brief A map between joint entity ids in the ECM to Joint Entities in
   /// ign-physics
   public: std::unordered_map<Entity, JointPtrType> entityJointMap;
+
+  /// \brief A map between model entity ids in the ECM to whether its battery
+  /// has drained.
+  public: std::unordered_map<Entity, bool> entityOffMap;
 
   /// \brief used to store whether physics objects have been created.
   public: bool initialized = false;
@@ -508,6 +514,15 @@ void PhysicsPrivate::CreatePhysicsEntities(const EntityComponentManager &_ecm)
         this->entityJointMap.insert(std::make_pair(_entity, jointPtrPhys));
         return true;
       });
+
+  _ecm.EachNew<components::BatterySoC>(
+      [&](const Entity & _entity, const components::BatterySoC *)->bool
+      {
+        // Parent entity of battery is model entity
+        this->entityOffMap.insert(std::make_pair(
+          _ecm.ParentEntity(_entity), false));
+        return true;
+      });
 }
 
 //////////////////////////////////////////////////
@@ -563,6 +578,17 @@ void PhysicsPrivate::RemovePhysicsEntities(const EntityComponentManager &_ecm)
 void PhysicsPrivate::UpdatePhysics(const EntityComponentManager &_ecm)
 {
   IGN_PROFILE("PhysicsPrivate::UpdatePhysics");
+  // Battery state
+  _ecm.Each<components::BatterySoC>(
+      [&](const Entity & _entity, const components::BatterySoC *_bat)
+      {
+        if (_bat->Data() <= 0)
+          entityOffMap[_ecm.ParentEntity(_entity)] = true;
+        else
+          entityOffMap[_ecm.ParentEntity(_entity)] = false;
+        return true;
+      });
+
   // Handle joint state
   _ecm.Each<components::Joint, components::Name>(
       [&](const Entity &_entity, const components::Joint *,
@@ -571,6 +597,20 @@ void PhysicsPrivate::UpdatePhysics(const EntityComponentManager &_ecm)
         auto jointIt = this->entityJointMap.find(_entity);
         if (jointIt == this->entityJointMap.end())
           return true;
+
+        // Model is out of battery
+        if (this->entityOffMap[_ecm.ParentEntity(_entity)])
+        {
+          std::size_t nDofs = jointIt->second->GetDegreesOfFreedom();
+          for (std::size_t i = 0; i < nDofs; ++i)
+          {
+            jointIt->second->SetForce(i, 0);
+            // TODO(anyone): Only for diff drive, which does not use
+            //   JointForceCmd. Remove when it does.
+            jointIt->second->SetVelocity(i, 0);
+          }
+          return true;
+        }
 
         auto force = _ecm.Component<components::JointForceCmd>(_entity);
         if (force)
@@ -593,16 +633,24 @@ void PhysicsPrivate::UpdatePhysics(const EntityComponentManager &_ecm)
         else
         {
           // Only set joint velocity if joint force is not set.
-          // TODO(addisu) We should use a different component for setting
-          // velocities. This component should be used to report the current
-          // velocity of the joint.
-          auto vel1 = _ecm.Component<components::JointVelocity>(_entity);
-          if (vel1)
-            jointIt->second->SetVelocity(0, vel1->Data());
-
-          auto vel2 = _ecm.Component<components::JointVelocity2>(_entity);
-          if (vel2)
-            jointIt->second->SetVelocity(1, vel2->Data());
+          auto velCmd = _ecm.Component<components::JointVelocityCmd>(_entity);
+          if (velCmd)
+          {
+            if (velCmd->Data().size() != jointIt->second->GetDegreesOfFreedom())
+            {
+              ignwarn << "There is a mismatch in the degrees of freedom between"
+                      << " Joint [" << _name->Data() << "(Entity=" << _entity
+                      << ")] and its JointVelocityCmd component. The joint has "
+                      << velCmd->Data().size() << " while the component has "
+                      << jointIt->second->GetDegreesOfFreedom() << ".\n";
+            }
+            std::size_t nDofs = std::min(velCmd->Data().size(),
+                jointIt->second->GetDegreesOfFreedom());
+            for (std::size_t i = 0; i < nDofs; ++i)
+            {
+              jointIt->second->SetVelocity(i, velCmd->Data()[i]);
+            }
+          }
         }
 
         return true;
@@ -888,11 +936,25 @@ void PhysicsPrivate::UpdateSim(EntityComponentManager &_ecm) const
         return true;
       });
 
-  // Clear pending Forces
+  // Clear pending commands
   _ecm.Each<components::JointForceCmd>(
       [&](const Entity &, components::JointForceCmd *_force) -> bool
       {
         std::fill(_force->Data().begin(), _force->Data().end(), 0.0);
+        return true;
+      });
+
+  _ecm.Each<components::ExternalWorldWrenchCmd >(
+      [&](const Entity &, components::ExternalWorldWrenchCmd *_wrench) -> bool
+      {
+        _wrench->Data().Clear();
+        return true;
+      });
+
+  _ecm.Each<components::JointVelocityCmd>(
+      [&](const Entity &, components::JointVelocityCmd *_vel) -> bool
+      {
+        std::fill(_vel->Data().begin(), _vel->Data().end(), 0.0);
         return true;
       });
 
@@ -914,6 +976,23 @@ void PhysicsPrivate::UpdateSim(EntityComponentManager &_ecm) const
         return true;
       });
 
+  // Update joint Velocities
+  _ecm.Each<components::Joint, components::JointVelocity>(
+      [&](const Entity &_entity, components::Joint *,
+          components::JointVelocity *_jointVel) -> bool
+      {
+        auto jointIt = this->entityJointMap.find(_entity);
+        if (jointIt != this->entityJointMap.end())
+        {
+          _jointVel->Data().resize(jointIt->second->GetDegreesOfFreedom());
+          for (std::size_t i = 0; i < jointIt->second->GetDegreesOfFreedom();
+               ++i)
+          {
+            _jointVel->Data()[i] = jointIt->second->GetVelocity(i);
+          }
+        }
+        return true;
+      });
   this->UpdateCollisions(_ecm);
 }